created admin portal

+++ src/app/services/user.service.ts
@@ -12,15 +12,15 @@ export class UserService {
   constructor(private http: HttpClient) { }
 
   getUser(username: string) {
-    if (username === 'admin') {
-      return this.http.get(`${baseUrl}/user/get/all`, {headers: this.setHeader()})
-    } else {
-      return this.http.get(`${baseUrl}/user/${username}`, {headers: this.setHeader()})
-    }
+    return this.http.get(`${baseUrl}/user/${username}`, {headers: this.setHeader()})
   }
 
   getUserById (id: any){
     return this.http.get(`${baseUrl}/user/get/${id}`, {headers: this.setHeader()})
+  }
+
+  getAllUsers() {
+    return this.http.get(`${baseUrl}/user/find/all`, {headers: this.setHeader()})
   }
 
   editUser(id: number, newData:any) {
@@ -28,7 +28,12 @@ export class UserService {
   }
 
   deleteUser(id: number) {
-    return this.http.delete(`${baseUrl}/user/${id}`, {headers: this.setHeader()})
+    if (id === 1) {
+      console.log('cannot delete admin')
+      return this.http.get(`${baseUrl}/user/get/${id}`, {headers: this.setHeader()})
+    } else {
+      return this.http.delete(`${baseUrl}/user/${id}`, {headers: this.setHeader()})
+    }
   }
 
   private setHeader(): HttpHeaders {
